Hopefully stop spurious candidate deletion

+++ src/main/java/nl/bioinf/cawarmerdam/compound_evolver/model/pipeline/ValidateConformersStep.java
@@ -93,7 +93,7 @@ public class ValidateConformersStep implements PipelineStep<Candidate, Candidate
                 System.out.println("Accepted candidate: " + candidate);
             }
             return candidate;
-        } else if (deleteInvalid && candidate.canBeDeleted()) {
+        } else if (deleteInvalid && candidate.canBeDeleted() && candidate.getScoredConformersFile() == null) {
             try {
                 FileUtils.deleteDirectory(candidate.getConformersFile().getParent().toFile());
             } catch (IOException exception) {
